+++ src/core/functools/not-implemented/index.ts
@@ -11,50 +11,41 @@
  * @packageDocumentation
  */
 
-import {
+import { warn, warned, WarnOptions, InlineWarnOptions, WarnedFn } from 'core/functools/warning';
 
-	NotImplementedOptions,
-	InlineNotImplementedOptions,
-	NotImplementedFn
-
-} from 'core/functools/not-implemented/interface';
-
-export * from 'core/functools/deprecation/interface';
-
-const
-	consoleCache = Object.create(null);
+export * from 'core/functools/warning/interface';
 
 /**
  * Marks the specified function as not implemented
  *
  * @param opts - additional options
- * @param [fn] - function for wrapping
+ * @param fn - function to wrap
  */
 export function notImplement<T extends Function>(
-	opts: NotImplementedOptions,
+	opts: WarnOptions,
 	fn: T
-): T extends ((...args: infer A) => infer R) ? NotImplementedFn<A, R> : T;
+): T extends ((...args: infer A) => infer R) ? WarnedFn<A, R> : T;
 
 /**
  * Emits a "not implemented" warning with the specified parameters
  * @param opts - additional options
  */
-export function notImplement(opts: InlineNotImplementedOptions): void;
+export function notImplement(opts: InlineWarnOptions): void;
 
 /**
  * Marks the specified function as not implemented
- * @param fn - function for wrapping
+ * @param fn - function to wrap
  */
 export function notImplement<T extends Function>(fn: T): T extends ((...args: infer A) => infer R) ?
-	NotImplementedFn<A, R> :
+	WarnedFn<A, R> :
 	T;
 
 export function notImplement<T extends Function>(
-	fnOrParams: NotImplementedOptions | InlineNotImplementedOptions | T,
+	fnOrParams: WarnOptions | InlineWarnOptions | T,
 	fn?: T
 ): T | undefined {
 	let
-		p: NotImplementedOptions;
+		p: WarnOptions;
 
 	if (Object.isSimpleFunction(fnOrParams)) {
 		fn = fnOrParams;
@@ -64,52 +55,7 @@ export function notImplement<T extends Function>(
 		p = fnOrParams;
 	}
 
-	if (!fn) {
-		wrapper();
-		return;
-	}
-
-	wrapper.notImplemented = p;
-	function wrapper(this: unknown, ...args: unknown[]): unknown {
-		//#unless isProd
-
-		const
-			name = p.name ?? fn?.name,
-			type = p.type ?? 'function';
-
-		const
-			msg = [`The ${type} "${name}" is not implemented.`];
-
-		if (p.alternative != null) {
-			if (Object.isString(p.alternative)) {
-				msg.push(`Please use "${p.alternative}" instead.`);
-
-			} else if (p.alternative.source != null) {
-				msg.push(`Please use "${p.alternative.name}" from "${p.alternative.source}" instead.`);
-
-			} else {
-				msg.push(`Please use "${p.alternative.name}" instead.`);
-			}
-		}
-
-		if (p.notice != null) {
-			msg.push(p.notice);
-		}
-
-		const
-			str = msg.join(' ');
-
-		if (consoleCache[str] == null) {
-			console.warn(str);
-			consoleCache[str] = true;
-		}
-
-		//#endunless
-
-		return fn?.apply(this, args);
-	}
-
-	return <any>wrapper;
+	return <any>warn({context: 'notImplemented', ...p}, <T>fn);
 }
 
 /**
@@ -147,40 +93,17 @@ export function notImplemented(target: object, key: string | symbol, descriptor:
  * }
  * ```
  */
-export function notImplemented(opts?: NotImplementedOptions): Function;
+export function notImplemented(opts?: WarnOptions): Function;
 
 export function notImplemented(
-	opts?: NotImplementedOptions | object,
+	opts?: WarnOptions | object,
 	key?: string | symbol,
 	descriptor?: PropertyDescriptor
 ): Function | undefined {
-	const f = (name, descriptor, opts?) => {
-		const
-			{get, set, value: method} = descriptor;
-
-		if (get != null) {
-			descriptor.get = notImplement({type: 'accessor', ...opts, name}, get);
-		}
-
-		if (set != null) {
-			descriptor.set = notImplement({type: 'accessor', ...opts, name}, set);
-		}
-
-		if (get != null || set != null) {
-			return;
-		}
-
-		if (!Object.isFunction(method)) {
-			throw new TypeError(`descriptor.value is not a function: ${method}`);
-		}
-
-		descriptor.value = notImplement({type: 'method', ...opts, name}, method);
-	};
-
 	if (arguments.length > 1) {
-		f(key, descriptor);
+		warned({context: 'notImplemented'})(key, descriptor);
 		return;
 	}
 
-	return (target, key, descriptor) => f(key, descriptor, opts);
+	return (target, key, descriptor) => warned({context: 'notImplemented', ...opts})(key, descriptor);
 }